fix frontend structure
